requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
SpeechRecognition>=3.10.0
google-generativeai>=0.4.0
keyboard>=0.13.5
rich>=13.7.0
python-dotenv>=1.0.0
